Extract RecentlyVisitedCellViewModel
